+++ pepper/rc/changelog.md
@@ -12,6 +12,7 @@
 - added css syntax
 - changed bracket objects to invert bracket positions if invoked with the closing bracket. that is, `a)` will now select text between `)` and `(` instead of `(` and `)`
 - changed `StatusBar` to `Logger` (and `editor.status_bar` to `editor.logger`)
+- changed `print` command to `log` command which accepts a `<log-kind>` parameter (use `log status <args...>` for old behavior)
 - added logging to the editor which you can open it with the `open-log` command
 
 # 0.27.0